+++ src/app/pages/settings/settings.component.ts
@@ -37,7 +37,7 @@ export class DatasourceAvailablePipe implements PipeTransform {
         </ng-template>
 
         <div class="app-root app-settings-page" responsive>
-            <mat-toolbar color="primary" class="mat-elevation-z2 app-toolbar" responsive>
+            <mat-toolbar color="primary" class="app-toolbar" responsive [class.mat-elevation-z2]="!vp.sm">
                 <div style="display: flex; align-items: center">
                     <button mat-icon-button (click)="back()">
                         <mat-icon style="color: var(--text-contrast)">close</mat-icon>